+++ x/feemarket/types/feemarket.pb.go
@@ -26,22 +26,22 @@ const _ = proto.GoGoProtoPackageIsVersion3 // please upgrade the proto package
 
 // Params defines the EVM module parameters
 type Params struct {
-	// no base fee forces the EIP-1559 base fee to 0 (needed for 0 price calls)
+	// no_base_fee forces the EIP-1559 base fee to 0 (needed for 0 price calls)
 	NoBaseFee bool `protobuf:"varint,1,opt,name=no_base_fee,json=noBaseFee,proto3" json:"no_base_fee,omitempty"`
-	// base fee change denominator bounds the amount the base fee can change
+	// base_fee_change_denominator bounds the amount the base fee can change
 	// between blocks.
 	BaseFeeChangeDenominator uint32 `protobuf:"varint,2,opt,name=base_fee_change_denominator,json=baseFeeChangeDenominator,proto3" json:"base_fee_change_denominator,omitempty"`
-	// elasticity multiplier bounds the maximum gas limit an EIP-1559 block may
+	// elasticity_multiplier bounds the maximum gas limit an EIP-1559 block may
 	// have.
 	ElasticityMultiplier uint32 `protobuf:"varint,3,opt,name=elasticity_multiplier,json=elasticityMultiplier,proto3" json:"elasticity_multiplier,omitempty"`
-	// height at which the base fee calculation is enabled.
+	// enable_height defines at which block height the base fee calculation is enabled.
 	EnableHeight int64 `protobuf:"varint,5,opt,name=enable_height,json=enableHeight,proto3" json:"enable_height,omitempty"`
-	// base fee for EIP-1559 blocks.
+	// base_fee for EIP-1559 blocks.
 	BaseFee github_com_cosmos_cosmos_sdk_types.Int `protobuf:"bytes,6,opt,name=base_fee,json=baseFee,proto3,customtype=github.com/cosmos/cosmos-sdk/types.Int" json:"base_fee"`
 	// min_gas_price defines the minimum gas price value for cosmos and eth transactions
 	MinGasPrice github_com_cosmos_cosmos_sdk_types.Dec `protobuf:"bytes,7,opt,name=min_gas_price,json=minGasPrice,proto3,customtype=github.com/cosmos/cosmos-sdk/types.Dec" json:"min_gas_price"`
-	// min gas denominator bounds the minimum gasUsed to be charged
-	// to senders based on GasLimit
+	// min_gas_multiplier bounds the minimum gas used to be charged
+	// to senders based on gas limit
 	MinGasMultiplier github_com_cosmos_cosmos_sdk_types.Dec `protobuf:"bytes,8,opt,name=min_gas_multiplier,json=minGasMultiplier,proto3,customtype=github.com/cosmos/cosmos-sdk/types.Dec" json:"min_gas_multiplier"`
 }